fix change and reset pw
