+++ src/shared/shared/shared.module.ts
@@ -1,7 +1,9 @@
-import { Module } from '@nestjs/common';
+import { forwardRef, Module } from '@nestjs/common';
+import { UserModule } from 'src/main/user/user.module';
 import { SharedService } from './shared.service';
 
 @Module({
+  imports: [forwardRef(() => UserModule)],
   providers: [SharedService],
   exports: [SharedService],
 })
