+++ src/shared/shared/shared.service.ts
@@ -1,7 +1,19 @@
-import { Injectable } from '@nestjs/common';
+import User from 'src/main/user/user.entity';
+import { UserService } from 'src/main/user/user.service';
+import {
+  Injectable,
+  BadRequestException,
+  forwardRef,
+  Inject,
+} from '@nestjs/common';
 import * as bcrypt from 'bcrypt';
+import { VerifyOTPDto } from 'src/main/auth/dto/verifyOTPDto';
 @Injectable()
 export class SharedService {
+  constructor(
+    @Inject(forwardRef(() => UserService))
+    private userService: UserService,
+  ) {}
   public async hashPassword(password: string): Promise<string> {
     const salt: string = await bcrypt.genSalt();
     return await bcrypt.hash(password, salt);
@@ -39,5 +51,24 @@ export class SharedService {
       code = Math.floor(Math.random() * 1000000);
     } while (code < 100000);
     return code;
+  }
+
+  public async verifyOTP(verifyOTPDto: VerifyOTPDto): Promise<void> {
+    const user: User = await this.userService.findByUsername(
+      verifyOTPDto.username,
+    );
+    if (!user) {
+      throw new BadRequestException('Not found username.!');
+    }
+    const diff = Math.abs(
+      Date.now() - user.phoneNumberVerifyCodeExpire.getTime(),
+    );
+    const minutes = Math.floor(diff / 1000 / 60);
+    if (verifyOTPDto.otp === user.phoneNumberVerifyCode && minutes > 5) {
+      throw new BadRequestException('OTP time up.!');
+    }
+    if (verifyOTPDto.otp !== user.phoneNumberVerifyCode && minutes < 5) {
+      throw new BadRequestException('OTP is wrong.!');
+    }
   }
 }
